Provide a pledge() implementation using waive()
This maps usage of pledge() to waive(), which itself uses seccomp filters to
restrict which things can a process do. The waive() implementation works by
passing to it the set of capabilities which the process resigns to use, while
pledge() receives the capabilities which are to be used, so the disjoint set
is calculated and then passed to waive().

Uses https://github.com/dimkr/libwaive