Always add flags from pkg-config to linker command line
The linker flags obtained from pkg-config may include flags other than "-lbsd"
such as a library search path ("-L...") for the linker, so putting the flags
from pkg-config in the list of dependencies for the "signify" target may not
work in some systems which place libraries in locations other than the usual
/lib, /usr/lib, or /usr/local/lib. This patch moves the flags obtained from
pkg-config to a different variable, $(LIBBSD_LDFLAGS), which used in the
linker command line but not added to the list of dependencies.

When using a bundled libbsd, the locally-built libbsd.a is still added to
the list of dependencies for the "signify" target, using the $(LIBBSD_DEPS)
variable. This will be empty when using pkg-config.

This fixes issue #6.