Remove unused arc4random.[hc] sources
As it turns out, the code has been using libbsd's version for a long
while already, and the bundled version of the arc4random functions was
not being used. This removes the files, and changes crypto_api.h to
include <bsd/stdio.h> in order to avoid having to forward-declare
functions in compat.h manually.

This is just a cleanup without any functional changes.