Automatically use glibc's explicit_bzero() where available
This adds build-time detection of the glibc version to enable using the
implementation of explicit_bzero() provided by glibc 2.25 and newer:

Release notes for glibc 2.25:

  https://sourceware.org/ml/libc-alpha/2017-02/msg00079.html

The bundled version is still used for non-glibc systems, and it is
possible to disable the automatic detection by setting the BZERO Make
variable (i.e. in the command line).